requirements.txt
asn1crypto==0.24.0
cachetools==3.1.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
future==0.17.1
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-images-download==2.8.0
Google-Images-Search==0.3.8
httplib2==0.13.0
idna==2.8
numpy==1.16.4
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pyfiglet==0.8.post1
python-resize-image==1.1.19
python-telegram-bot==11.1.0
requests==2.22.0
rsa==4.0
selenium==3.141.0
six==1.12.0
termcolor==1.1.0
uritemplate==3.0.0
urllib3==1.25.3
